+++ database/migrations/2019_04_07_115306_create_products_table.php
@@ -14,13 +14,13 @@ class CreateProductsTable extends Migration
     public function up()
     {
         Schema::create('products', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('system_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('system_id')->unsigned();
             $table->foreign('system_id')->references('id')->on('system_type');
             $table->integer('product_id')->unique();
-            $table->integer('venue_id')->unsigned();
+            $table->bigInteger('venue_id')->unsigned();
             $table->foreign('venue_id')->references('id')->on('venues');
-            $table->integer('merchant_id')->unsigned();
+            $table->bigInteger('merchant_id')->unsigned();
             $table->foreign('merchant_id')->references('id')->on('merchants');
             $table->string('product_brand_ids')->nullable();
             $table->string('product_category_ids')->nullable();
